data reformatting continued
